# Skip rendering on low-end devices