Check if we missed a game tick.
This can happen when we run a low-end device that can't render fast enough (typically 20fps).
If this happens, try to speed-up the game by skipping the rendering.
This avoids desyncs and hourglasses when running multiplayer and slowdowns in singleplayer.

Edge-case:
If the game runs so slow, that we can't calculate the gamelogic in 50ms (one game tick) no frames will be rendered.
But if this happens the game should be unplayable anyway.

<details><summary>Comparison video</summary>

Comparison video with a ~10 fps (for testing I added a `_sleep(100);` in `DrawAndBlit`).

## Master ~10fps

https://user-images.githubusercontent.com/25415264/219756349-01cbef34-8cf7-41b7-8748-116a8f994ed3.mp4

## PR ~10fps

https://user-images.githubusercontent.com/25415264/219756458-23f09e44-0cd6-41ff-95fb-41d1aecaa30b.mp4

## Fast devices ~160fps

https://user-images.githubusercontent.com/25415264/219756974-4753f33d-5ceb-4fd3-99ca-45c537961542.mp4

</details>